feat(RootLayout.tsx): create RootLayout component to replace Root component
